% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_recebidos.R
\name{doclink}
\alias{doclink}
\title{doclink}
\usage{
doclink(docname, campos)
}
\arguments{
\item{docname}{document name}

\item{campos}{xml nodes for columns}
}
\value{
url string
}
\description{
doclink
}
