% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\docType{data}
\name{docs}
\alias{docs}
\title{Documentos recebidos}
\format{
Tabela com 7 colunas:
\describe{
\item{n}{número}
\item{arquivo}{lista de arquivos}
\item{data_recebimento}{data de recebimento}
\item{remetente}{remetente}
\item{origem}{origem}
\item{descricao}{descrição}
\item{caixa}{caixa}
}
}
\usage{
docs
}
\description{
Informações sobre documentos recebidos na CPI da Pandemia
}
\keyword{datasets}
