% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_doclink}
\alias{get_doclink}
\title{get_doclink}
\usage{
get_doclink(docname, campos)
}
\arguments{
\item{docname}{document name}

\item{campos}{xml nodes for columns}
}
\value{
url string
}
\description{
get_doclink
}
