% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_recebidos.R, R/scraper.R
\name{parse_row}
\alias{parse_row}
\title{parse_row}
\usage{
parse_row(row)

parse_row(row)
}
\arguments{
\item{row}{xml node for row}
}
\value{
one row tibble

one row tibble
}
\description{
parse_row

parse_row
}
