% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{req_apreciados}
\alias{req_apreciados}
\title{req_apreciados}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 800 rows and 6 columns.
}
\usage{
req_apreciados
}
\description{
req_apreciados
}
\keyword{datasets}
