% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_requeridos.R
\name{table_reqs}
\alias{table_reqs}
\title{Tabela de requerimentos}
\usage{
table_reqs(apreciado = TRUE)
}
\arguments{
\item{apreciado}{requerimento apreciado ou não (TRUE, FALSE)}
}
\value{
tibble
}
\description{
Tabela de requerimentos
}
